+++ CHANGES.rst
@@ -1,5 +1,25 @@
 Changelog
 =========
+6.1.2 (unreleased 2025-03-19)
+------------------
+
+Minor changes:
+
+- Set default value for ``params`` as ``params={}`` in mulitple constructors in ``icalendar.prop`` to improve usability.
+- Improve object initialization performance in ``icalendar.prop``.
+- Add type hint for ``params`` in multiple constructors in ``icalendar.prop``.
+
+Breaking changes:
+
+- ...
+
+New features:
+
+- ...
+
+Bug fixes:
+
+- ...
 
 6.1.2 (unreleased)
 ------------------